requirements.txt
Django==1.5.11
Markdown==2.5.2
MySQL-python==1.2.5
Pygments==2.0.1
argparse==1.2.1
django-jquery==1.9.1
django-pagedown==0.1.0
django-pagination==1.0.7
django-simple-menu==1.0.9
django-treemenus==0.9.2
docutils==0.12
python-memcached==1.53
wsgiref==0.1.2
